+++ src/indra_cogex/client/enrichment/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+"""A collection of implementations of gene set enrichment analysis methodologies."""
